+++ lib/lerk/lerk.rb
@@ -18,6 +18,7 @@ module Lerk
 
       @bot = Discordrb::Commands::CommandBot.new(token: @token, client_id: @client_id, prefix: @prefix)
 
+      create_rate_limiters
       bind_commands
     end
 
@@ -32,8 +33,47 @@ module Lerk
     private
     def bind_commands
       @bot.command :hive2 do |event, steam_id|
+        # Per-use rate limit
+         if @rate_limiter.rate_limited?(:hive2_user_api_calls, event.author)
+           puts "Hit rate limit for #{ event.author.username }, throttling..."
+           msg = <<EOF
+Hello there!
+
+You have reached the rate limit of Hive 2 queries.
+Please wait a few seconds before you issue another query. :)
+EOF
+           event.author.pm msg
+           next
+         end
+
+        # Global rate limit
+        if @rate_limiter.rate_limited?(:hive2_global_api_calls, :hive_get_player_data)
+          puts 'Hit global rate limit, throttling...'
+          next
+        end
+
         cmd_hive2(event, steam_id)
       end
+
+      @bot.command :hi do |event|
+        event.author.pm 'Hai'
+      end
+    end
+
+    def create_rate_limiters
+      @rate_limiter = Discordrb::Commands::SimpleRateLimiter.new
+      # Global rate limit.
+      @rate_limiter.bucket(
+        :hive2_global_api_calls,
+        limit:     Config::HIVE2_GLOBAL_RATE_LIMIT,
+        time_span: Config::HIVE2_GLOBAL_RATE_TIME_SPAN
+      )
+
+      @rate_limiter.bucket(
+        :hive2_user_api_calls,
+        limit:     Config::HIVE2_USER_RATE_LIMIT,
+        time_span: Config::HIVE2_USER_RATE_TIME_SPAN
+      )
     end
 
     def cmd_hive2(event, steam_id)